backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
requests==2.32.3
httpx==0.27.2
gitpython==3.1.43
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
jinja2==3.1.4
markdown==3.7
streamlit==1.38.0
pandas==2.2.2
plotly==5.24.1
openai==1.47.1
pytest==8.3.3
